Move login loic into main window, wire up login button
